+++ quickbooks/client.py
@@ -45,8 +45,8 @@ class QuickBooks(object):
     _BUSINESS_OBJECTS = [
         "Account", "Attachable", "Bill", "BillPayment",
         "Class", "CreditMemo", "Customer", "CompanyCurrency",
-        "Department", "Deposit", "Employee", "Estimate", "Invoice",
-        "Item", "JournalEntry", "Payment", "PaymentMethod",
+        "Department", "Deposit", "Employee", "Estimate", "ExchangeRate", "Invoice",
+        "Item", "JournalEntry", "Payment", "PaymentMethod", "Preferences",
         "Purchase", "PurchaseOrder", "RefundReceipt",
         "SalesReceipt", "TaxAgency", "TaxCode", "TaxService/Taxcode", "TaxRate", "Term",
         "TimeActivity", "Transfer", "Vendor", "VendorCredit", "CreditCardPayment",