+++ tests/integration/test_base.py
@@ -17,7 +17,7 @@ class QuickbooksTestCase(TestCase):
         )
 
         self.qb_client = QuickBooks(
-            minorversion=54,
+            minorversion=59,
             auth_client=self.auth_client,
             refresh_token=os.environ.get('REFRESH_TOKEN'),
             company_id=os.environ.get('COMPANY_ID'),